Revert "Llama7b"
